gcc 7: fix crashes due to corruption of xmm0 register
This patch fixes #890, which reported multiple cases of early crashes of OSv
when compiled with gcc 7. The common feature of all these crashes was that
a small random part of memory got zeroed, with various sad consequences.

The cause of these crashes is failing to correctly save FPU state. In modern
code the FPU (and SSE et al.) is used not just for floating point operations,
but also for various vector operations generated by the compiler as an
alternative to multiple instructions. Corrupting the FPU state can cause
such vector operations to go haywire, including zeroing the wrong place or
length of memory memory.

In some places, we save the FPU state on the current stack, using the
"fpu_lock" class. That class's constructor calls fpu_state_init() to
initialize the save area, and then fpu_state_save()/fpu_state_restore()
to save and restore the state.

Because fpu_state_init() is called before saving the FPU state, it must not
modify any of the FPU registers. But unfortunately, in gcc 7, it did: With
O2 optimization enabled, it zeroed the xmm0 register:

The XSAVE-based variant of fpu_state_init() (fpu_state_init_xsave()) wants
to zero 24 bytes of memory. It calls memset(), but the compiler optimizes
this call away and replaces it with direct instructions to zero 24 bytes.
Prior to gcc 7, this generated 3 MOVQ instructions, totalling 33 bytes of
machine code:

     48 c7 87 00 02 00 00    movq   $0x0,0x200(%rdi)
     00 00 00 00
     48 c7 87 08 02 00 00    movq   $0x0,0x208(%rdi)
     00 00 00 00
     48 c7 87 10 02 00 00    movq   $0x0,0x210(%rdi)
     00 00 00 00

starting with gcc 7, a significantly more elaborate technique using SSE
registers is used to zero the first 16 bytes. The result is only 22 bytes
of code, saving a "whopping" 11 bytes of code:

     66 0f ef c0             pxor   %xmm0,%xmm0
     48 c7 87 10 02 00 00    movq   $0x0,0x210(%rdi)
     00 00 00 00
     0f 11 87 00 02 00 00    movups %xmm0,0x200(%rdi)

In the case of fpu_state_init_xsave(), we of course do not care about
saving of 11 bytes of code, but the function is ruined by using SSE
registers. So this patch adds a tag on this function telling gcc to not
use SSE when compiling this function.

P.S. This patch has to be a new world record. Zero lines of code,
and 50 lines of explanations :-)

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170723195843.23951-1-nyh@scylladb.com>