gcc 7: fix crash when initial-exec TLS is used
When the initial-exec TLS model is used, as in our tst-tls.so, the
application's TLS area moves OSv's own TLS area by an offset of
"initial_tls_size" which wasn't aligned with any particular alignment.

But the linker knows (see arch/x64/loader.ld) that the TLS template segment
is aligned at a 64-byte boundary, and objects placed there were placed with
a particular alignment. For example, the thread_signal_mask TLS variable is
set (by sigprocmask()) for every new pthread, and because this is a 128-bit
variable, Gcc, starting with version 7, uses the "MOVAPS" SSE instruction
to set it. This instruction assumes that this variable is 16-byte aligned,
as was explicitly specified for this variable. But if initial_tls_size is
8 (mod 16), as it is in the "tst-tls.so" test, the thread_signal_mask TLS
variable is no longer 16-byte aligned, and the test crashes with a #GP fault.

So this patch fixes the initial-exec initial_tls_size to be 64-byte aligned
so not to break the alignment of OSv's own TLS variables. Also, to be on
the safe size, we use aligned_alloc to do the actual allocation, although
this is not strictly necessary (all allocations larger than a page are
page-aligned in any case).

Fixes #894.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170724080544.29220-1-nyh@scylladb.com>