libc: add functions required for libpython2.7.so
When we try to reuse python from the build system (instead of
compiling it from source), libpython2.7.so requires additional
libc/glibc symbols. It seems that (at least for most trivial python
programs) those functions are not really called and can be implemented
as stubs. The patch was prepared on Fedora 25.

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>
Message-Id: <20170801161142.31048-1-justin.cinkelj@xlab.si>