scripts/build: make it slightly harder to remove all your files :-)
The new "export=" and "export_dir=" parameters delete the entire content
of the target directory before exporting to it. I am not sure I like this
personally, but what I'd like even less is the deletion of the wrong
directory :-)

In the existing code, if the current directory or an explictly given
export_dir contains a space in its name, it will pass only the part up
to the first space as the "-e" option, and cause that unindented path
name to be deleted, potentially (although unlikely) deleting an unrelated
existing directory.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>