ELF: fix assertion failure when trying to run short non-executable
When trying to run a short non-executable (e.g. "scripts/run.py -e /etc/hosts")
we used to fail with an assertion failure in the function read(fileref, ...)
from fs/fs.cc. This function assumes it *can* read the desired buffer, and
if it can't (because the file is too short), it crashes. So when we try to
read the executable's header in a file which is shorter than the expected
size of such header, we crashed.

The solution is simple - verify that the size of the file is big enough
before calling the read function. If it is not, throw an exception.

This exception (and all the other exceptions in core/elf.cc during loading
an object) should be converted to a new type, as explained in issue #904.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>