Waldek Kozaczuk discovered that nginx, using sendfile(), is experiencing various failures when multiple concurrent connections are attempted.
It turns out that sendfile() had the wrong return type: Should have been
ssize_t (long), was int. So when sendfile() returned a -1, the application
sometimes thought it was a positive 4294967295. Of particular interest to
the nginx case, sendfile() is expected to temporarily fail when a non-
blocking socket's buffer is full, and this is "error" path is part of the
normal nginx operation, and did not happen correctly.

The fix is of course to change sendfile()'s return type to ssize_t, to match
Linux. Also in this patch we #include sys/sendfile.h to have the compiler
check the function's signature (if we had this #include earlier, we would
have caught this bug immediately).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>