Fix crash in read_partition_table()
When we find a disk device, we try to read its partition table to perhaps
divide the device into several partitions. If we fail to find this
partition table, it is fine - it means the disk has no partitions.
However, if we had an I/O error while reading the partition table
(as happens in the case of scsi-cd, see issue #909), the exiting
code continued to process the read data even though there was none,
and crashed on this processing or on the brelse() at its end.

In this case, the read_partition_table() should just do nothing, and
certainly not crash.

Fixes #903.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>