scsi: fix code typo
SCSI commands have two separate fields, "response" and "status", and the
code fixed by this patch had an obvious typo where the response was
copied into the status. Fixing this doesn't make any difference in
anything I was able to test (usually both are zero anyway), but I think
we should fix it, it might make a difference in some error situation
we will encounter in the future.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>