new object loading exceptions in core/elf.cc
Differentiate object loading errors from other loader errors.

Adds a new exception invalid_elf_error which is caught in
osv_execve so that it can set the error code to ENOEXEC.

Fixes #904.

One thing I should probably to explicitly point out is that there are a couple of runtime_errors in elf.cc. These are for memory_image::read() which is unsupported and for symbol lookups that fail. The latter raises the
question, is that an invalid elf?  I've chosen the answer "no" since
the caller has probably asked for the wrong symbol. In the end, I can
see both sides as it could be an invalid elf.

Signed-off-by: Brendan Cox <bcox@elotl.co>
Message-Id: <2ff01606-0234-42f8-aa27-c46911d7741f@googlegroups.com>