scripts/build: clearer message and no hang in case of missing file
Before this patch, if one of the files in the manifest is missing,
scripts/build prints the problem (as a long python stack trace) but this
message is mixed with the guest's output (of which files were saved in the
image), and the guest does not exit. This looks to the user like a silent
"hang" in the build (the error message is often not even visible - the user
has to scroll up to see it).

With this patch, when we have a problem uploading a file, we stop printing
the guest's output, print a clear message about the failure and stop the
guest. scripts/build will then say the build failed, and the reason will
be clear. And no hang.

Fixes #925

Signed-off-by: Nadav Har'El <nyh@scylladb.com>