Fix build on existing working directory after system upgrade
After commit c446634, the "httpserver" module no longer has a Makefile
and nothing is built in that directory - it is just a front for other
modules. However, the directory may have old files, notably usr.manifest,
which can break a build after the system was upgraded because some of the
old libraries it refers to no longer exist.
We need to add a Makefile with a "clean" target so at least "scripts/build
clean" will clean up this mess.

Fixes #931

Signed-off-by: Nadav Har'El <nyh@scylladb.com>