Be more explicit when build process fails to run qemu
OSv's build process (scripts/build) needs to run OSv in qemu, to upload files
to the new image. Currently, if qemu fails, run.py didn't notice and didn't
exit with an error code, and even if it did - upload_manifest.py which calls
it also didn't check run.py's error code. The result is that "scripts/build"
can fail to upload any files to the image, but be completely silent about it
and appears to have correctly created an image - it just doesn't have any
files.

This patch fixes both oversights: run.py now exit with an error code if qemu
fails, and upload_manifest.py exits with an error code if the run.py that it
called had an error. scripts/build already catches the error from
upload_manifest.py, so the result is that the build stops, ending with
a message about qemu having failed.

As an example why qemu can fail is that I had port 10000 taken by some
process, and upload_manifest.py to do port-forwarding to that port number
and couldn't.

In the future, we shoud choose a port number by random (like in commit
464f4e05c70e756a1a54e3751c710393371b7f5f) and verify it's not in use
because calling qemu. But until we do that, let's at least give the
user decent error messages.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>