memory: respect _bitset length bounds in page_range_allocator
The page_range_allocator attempts to merge neighbor page ranges
when a page range is freed in order to minimize memory fragmentation.

However, the original implementation didn't validate indexes before
checking the _bitset to determine whether a page range had a neighbor
or not.  This change make sure that indexes stay within bounds.

This is the correct fix for the fix attempted in
44f8324e3f2e53f55022a1d93c8b072514e7ac07

Signed-off-by: Timmons C. Player <timmons.player@spirent.com>
Message-Id: <1516038733-31578-1-git-send-email-timmons.player@spirent.com>