scripts/build: fix upload of unstrip-able shared objects
Before this patch, the upload process of scripts/build, when it sees
a shared object (*.so), it is stripped before uploading - into a new
file "*-stripped.so". If the strip process fails for some reason (e.g.,
the host's "strip" tool doesn't support a newer shared object), the
"*-stripped.so" file isn't created, but we nevertheless try to upload
it (and of course fail).

With this patch, if the strip fails, we print a warning message but
continue with uploading the original, unstripped, file.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>