Clean mutex_init()/mutex_destroy() warnings in gcc 8
In theory, the functions mutex_init()/mutex_destroy() are not needed at all
in C++ code, because creating a mutex object already initializes (zeros) it,
and it is destructed (doing nothing) automatically too. However, for historic
reasons, in a lot of code that started out as C and later moved to C++,
we do have calls to those functions in C++ code.

But, mutex_init() uses memset(), and Gcc starting in version 8 complains
about code which does memset and friends to a non-trivial C++ object.

So our fix is 3-pronged:

1. Reduce the usage of mutex_init()/mutex_destroy() in C++ code - e.g., if
   we just "new"ed a mutex, or it is a static variable, we don't need to
   initialize again.

2. For the little code remaining which was too messy to fix, and still
   calls mutex_init()/mutex_destroy() we leave these functions - but as
   *macros* instead of functions. This allows the compiler to complain
   only for the call site - not in every file which #includes mutex.h.

3. For the few mutex_init() call sites which remain and generate the gcc 8
   warnings, we will add (in a following patch) -Wno-class-memaccess
   to tell gcc to not test for this in those specific files.

Fixes #957

Signed-off-by: Nadav Har'El <nyh@scylladb.com>