Add dlvsym() function
The dlvsym() is a Linux-specific extension to the posix dlsym() which asks
for a specific version of the given symbol. It is needed by recent versions
of the openjdk8-fedora package (which always picks up the latest build from
Fedora, and changes frequently).

This is a stub implementation, which *ignores* the version string, and
prints a warning about this. From the little I tested, this seems to be
good enough to run Java. We can always complete the implementation later.

Fixes #961

Signed-off-by: Nadav Har'El <nyh@scylladb.com>