Fix strncpy() warnings in gcc 8
Gcc 8 added checks that when strncpy() is used to write to an array
of known length, this same length is not used as the "n" parameter of
strncpy - because if the same length is used, the result will not be
null terminated.

This patch fixes a few cases of this problem which gcc 8 found. The
solution is usually to use strlcpy() instead of strncpy().

Signed-off-by: Nadav Har'El <nyh@scylladb.com>