Fix warning in route.cc
In route.cc we have code which writes some socket addresses to a routing
socket. The write is supposed to be 8-byte aligned length, and the code
also *reads* a multiple of 8 bytes, even though the actual structure may
be shorter. Gcc 8 discovers this inconsistency (reading 56 bytes from a
structure known to be 54 bytes long) and complains.

The fix is to size the length of the copy correctly, and only 8-byte-align
the cursor of the write.

Other than quieting the compiler, this change shouldn't fix any real bug.

Fixes #958

Signed-off-by: Nadav Har'El <nyh@scylladb.com>