Avoid gcc 8 warnings on bzero()/memcpy() in old-style code
A lot of the BSD code used to be C code, which commonly bzero()ed or
memcpy()ed objects. In C++, this should not be done (objects have
constructors and assignments), and gcc 8 starts to warn about it.
Instead of fixing all these occurances, let's ask gcc to ignore this
warning on all code in the BSD directory. At least for now.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>