Fix warning on copying polymorphic type
It is rarely a good idea to capture an exception by value (making a copy)
instead of by reference, and it's even worse when the type is polymorphic
and copying it can result in "slicing". Gcc 8 started complaining about
such cases, even in cases where it doesn't matter (i.e., we don't even
use the exception we captured). So we need to change them to capture by
reference.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>