Fix warning about strncat
The implementation of __strcat_chk used strncat() in a way (with length
equal to the length of the source string) that causes Gcc 8 to warn about.
I think this warning is a false-positive, but nevertheless, the easiest
way to shut it up is to switch to strcat(), which works just as well here.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>