ELF: improve recognition of objects asking for on-load resolution
By default, we do lazy function call resolutions - functions are only looked
up on first use. But a shared object may ask (for various reasons) to do
all the resolutions up-front when the object is loaded. This is marked by
creating this object with the "-z now" linker option.

There are several standard ways a linker can use to mark this option in
the ELF file, though. The only method I encountered in the past was the
inclusion of a DT_BIND_NOW dynamic tag. However, it now appears that
some newer linkers do not do this, and instead put the bit DF_BIND_NOW
in the DT_FLAGS tag. Some documents claim that the DT_BIND_NOW tag is
obsolete. Also, some Linux linkers also put the DF_1_NOW flag in the
DT_FLAGS_1 tag, which was originally the Solaris-only way of marking this.

This patch adds support for all these alternative tags and flags. If any
one of those appears on the shared object, it is treated as bind-now.

Fixes #967.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180509130529.21156-1-nyh@scylladb.com>