gdb: fix enum lookup for gdb 8.1
This patch fixes a problem for gdb 8.1 to find the numeric value
of one enumerated types. This appears to me to be a gdb regression,
it is perhaps confused by the name "expired" being used as both a
function's name and an enumerated value. But nevertheless, since we
can easily work around this problem, and the new code works also
in older gdb, let's do it.

Fixes #964.
After this patch, "make check" finally passes on Fedora 28.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>