epoll: don't pass unknown bits to poll()
Our epoll implementation is based on poll(). We should only pass to poll
the event bits we know are identical between epoll and poll(). In particular,
EPOLLET and EPOLLONESHOT shouldn't be passed to poll().

Before this patch we passed EPOLLET and EPOLLONESHOT on to poll(), which
could confuse poll() inplementations, and in particular poll_no_poll()
which caused an assertion failure.

Refs #971.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>