poll: don't return POLLNVAL from poll_no_poll()
The trivial poll_no_poll() returned POLLNVAL if it got a bit it doesn't
recognize. This is wrong: poll() should ignore bits it doesn't recognize
and poll(1) states that POLLNVAL is only to be returned if the file
descriptor isn't open. This patch just drops the wrong test.

Before this patch, adding an extra bit not listed, like EPOLLRDHUP,
caused POLLNVAL to be returned. Worse, when using epoll, it did not
expect to ever get back POLLNVAL from poll(), and caused an assertion
failure.

Fixes #971.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>