epoll: test polling of /proc file
This test reproduces issue #971, and verifies that we fixed it in the
previous patches.

It epolls a file from /proc - Linux doesn't allow epolling an ext4 file,
but does allow to do that on /proc files. It then verifies that epoll
doesn't crash even with unorthodox request bits (as happened in #971).

The test can now be run on both Linux and OSv, and succeds on both.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>