epoll: fix regression in EPOLLET on network sockets
9eb6f2e38dc0103bd0d64a17221e2cc59b8cb138 changed epoll() not to pass "unknown"
bits to the file implementation's poll() method. This had a worthy cause:
issue #971 was caused by unknown bits which got to poll_no_poll() which in
turn complained about them.

However, this change caused a new regression in EPOLLET on sockets (see
issue #974 reporting this regression). The socket's poll function relies
on knowing whether EPOLLET is being used, because it needs to avoid an
optimization which would cause missed wakeups with EPOLLET (see commit
d41d748f89cf6e47820e8585cef6eedb837c05be for more information). So although
the EPOLLET bit is *not* traditionally supported by the poll() method, we
in OSv we do need to pass it on to poll(). Other bits, namely EPOLLONESHOT,
we still don't need to pass.

Fixes #974.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20180519182659.15445-1-nyh@scylladb.com>