Allow execution of non-public Java classes
OSv uses thin Java wrapper to bootstrap execution of JVM programs. It
does so by using reflection to select static main method of the program
entry class.

Java specification does not require the entry main class
to be public if program is started by standard java executable. However
on OSv we use our own java.so and thin Java wrapper (the io.osv.Jvm class)
which fails to execute the main method if the program class is not public.
It happens because Jvm class uses reflection and by default Java security
manager enforces that this class has access to the programs class being invoked
which it does NOT when the program class is non-public
(one example is latest Elasticsearch).

This patch supresses Java security access check to allow execution
of main method on non-public classes.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>