aarch64: add missing __NR_getrandom
Add missing __NR_getrandom definition to unbreak the aarch64 build.
The correct number is 278, because according to experts I found on
the Internet (;-)) in ARM64, unlike ARM, the system call numbers are
the ones from asm-generic.h.

It seems, by the way, that most of the other numbers in
include/api/aarch64/bits/syscall.h are wrong :-(

Signed-off-by: Nadav Har'El <nyh@scylladb.com>