Implement almost-in-place kernel decompression
This patch implements almost-in-place decompression of kernel as
described by #985 with caveat that the segments are decompressed
starting with the last one. More detailed explanation of how
this new decompression strategy works can be found in extensive
comments added to fastlz source code.

This new decompression strategy has following benefits:
- minimizes amount of RAM to execute OSv images (the new minimum
  RAM to execute native example with ROFS is 27.5 MB instead of 31MB).
  The difference is almost twofold when executing ramfs images with
  java-example.
- one no longer needs to adjust lzkernel_base manually to
  build bigger ramfs images.

Fixes #985

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Reviewed-by: Nadav Har'El <nyh@scylladb.com>