Clean build tree when doing scripts/build check or make check
Apparently there are missing dependencies in our Makefile
that in some cases lead to situation when certain
artifacts are not properly rebuilt and manual cleanup
of the build machine is necessary.

This patch changes the scripts/build to automatically
clean main build artifacts as well as modules before
build starts.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>