Fix assertion failure when running shared object linked with Gold
The "GNU_RELRO" header marks an area of memory which should be make
read-only (RO) after relocations (REL) have taken place.

Since headers describe an area of memory, and have their own permission
flags, it was natural to assume that RELRO would come with a "R" (read-only
flag). And this was indeed the case when linking with Gnu's classic linker
(ld.bfd). However, with the newer linker, Gold, the "RW" flags are used on
this header. This should make no difference: glibc's ld.so uses this
header to unconditionally make the memory read-only, ignoring the flags
(see _dl_protect_relro, glibc's elf/dl-reloc.c). We should do the same,
and drop the undue assert().

We saw this problem in the past, in issue #891, and switched in Makefile
to link with ld.bfd. However, some of the files compiled in Makefile,
such as mkfs.so, is compiled directly with gcc, not with separate
compilation and linking stages, so can end up being linked with ld.gold.

Refs #891.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>