Implement more space-efficient early memory allocation scheme
This patch implements a naive but efficient scheme of allocating
memory before SMP is enabled and regular malloc pools are setup.
Before this patch every single pre-SMP allocation would
be backed by full 4K page regardless if requested size
was 8 or 2000 bytes.

New pre-SMP allocation scheme is based on the fact that
relatively few objects are allocated (~ 2,000) and rarely
freed (< 20%) and most are very small (<50 bytes).
The algorithm is as simple as finding next big enough aligned
space in a 4K page of memory, marking its size and resetting
offset to the next free area of the page otherwise allocating
new 4K page. For details read comments in the code.

Even though this scheme is very simplistic it is
roughly 35 times more space efficient and allows
us to save around 6M of memory. It was measured that the net sum of
sizes (as is) of all requested allocations is typically around 125K
and new scheme uses ~44 pages (176K) of memory to satisfy all requests.

With this patch it is possible to run native example
with 18.5M of memory:
./scripts/build -j6 fs=rofs image=native-example
./scripts/run.py -c 1 -m 19M

Fixes #270

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20181211145428.28889-1-jwkozaczuk@gmail.com>