Delete assigned_virtio driver
This patch deletes all code related to supporting assigned
virtio driver functionality. The idea with the "assigned virtio" API
was a new API to let the user-space application interact directly with
the virtio ring, without kernel involvement. This would allow a user-space
TCP/IP implementation bypass OSv's TCP/IP implementation. The only
app using this API was Seastar, but Seastar cannot run on OSv anymore
and nobody else maintains this API. Given that this API cannot be
tested without Seastar and it gets in a way of upgrading virtio
layer to support modern PCI devices and virtio MMIO devices,
we have decided to remove it from OSv codebase.

In future if somebody finds it useful again we will bring it back
and adjust accordingly.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>