Make OSv boot without ACPI present
This patch makes OSv boot without requiring ACPI
to be present which for example is the case on firecracker.

We simply treat failure to find ACPI root pointer
as an indicator that ACPI is not available and
mark it off as such. Also if ACPI is off we power OSv off
using non-ACPI method and skip probing panic driver as it
relies on ACPI as well.

Lastly we skip MADT table parsing if APCI is off and
for now assume there is single vCPU only. Eventually
we should parse vCPU information from MP table as
an alternative.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>