Fix virtio-blk disk size reporting
This patch tweaks virtio blk disk size calculation
by using 512 as the size of the sector instead of
_config.blk_size which might not be even negotiated.

Per virtio 1.0 spec, 5.2.5, step 2:
"if the VIRTIO_BLK_F_BLK_SIZE feature is negotiated, blk_size can be read to determine the optimal sector size for the driver to use. This does not affect the units used in the protocol (always 512 bytes), but awareness of the correct value can affect performance."

This makes disk reporting work correctly on firecracker.

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>