build: don't fail build if pathname has space
Ideally, the build shouldn't care what is the pathname of the build
directory. But some parts of the build process use, for historical
reasons and for convenience, full pathnames, and some of that code did
not correctly handle the case where the pathname might contain spaces,
and the build fails.

The fix is to use relative pathnames where it's easy, but where it's not,
be more careful to quote the filename or avoid using the shell's argument
splitting in Python code.

Fixes #1034 for basic "make" and "scripts/build image=rogue", but some
modules (e.g., Java) still don't build properly in this case and should
be fixed separately.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20190314132451.26548-1-nyh@scylladb.com>