Simplify building images out of artifacts found on host filesystem
Building OSv images with apps that have to built from source from scratch is not fun
at best. With the advancement of ASLR more and more executables on
modern ditributions of Linux come as PIEs (Position Independent Executables)
which OSv mostly supports.

This patch provides new script aiming to simplify
building images out of the ELF files found on Linux host
filesystem.

The new script relies on lddtree that is part of pax-utils package.

The examples of using the script:
  ./scripts/manifest_from_host.sh ls                 # Create manifest for 'ls' executable
  ./scripts/manifest_from_host.sh -r /some/directory # Create manifest out of file in the directory
  ./scripts/manifest_from_host.sh -l libz.so.1       # Create manifest for libz.so.1 library
  ./scripts/manifest_from_host.sh -w ls && \
  ./script/build --append-manifest                   # Create manifest for 'ls' executable and build image

Signed-off-by: Waldemar Kozaczuk <jwkozaczuk@gmail.com>
Message-Id: <20190510035335.8178-1-jwkozaczuk@gmail.com>